Add gameobject prefab instantiation

+++ game-engine/build/clean.js
@@ -1,12 +1,20 @@
 const path = require("path");
 const rimraf = require("rimraf");
+const cc = require("./console-colors");
+const fs = require("fs");
 
 [
     path.resolve(__dirname, "../lib"),
     path.resolve(__dirname, "../lib-esm"),
     path.resolve(__dirname, "_bundles")
 ].forEach(path => {
-    rimraf(path, () => {
-        console.log("\x1b[31m", `Removed folder ${path}`, "\x1b[0m");
-    });
+    const timeStamp = new Date(Date.now()).toLocaleTimeString();
+
+    if (fs.existsSync(path)) {
+        rimraf(path, () => {
+            console.log(cc.fg.yellow, `${timeStamp} - Removed folder ${path}`, cc.reset);
+        });
+    } else {
+        console.log(`${timeStamp} - Ignoring ${path} - folder not found`);
+    }
 });
